[buckets-exploration] - Add navigable header to buckets
