+++ src/components/Dashboard/components/DashboardContent/styles.scss
@@ -2,15 +2,4 @@
 
 .dashboard-content {
   height: 100%;
-
-  .month-header {
-    margin-top: $section_header_vertical-margin;
-    margin-bottom: $section_header_vertical-margin;
-    align-items: center;
-
-    div:nth-child(3) {
-      display: flex;
-      justify-content: flex-end;
-    }
-  }
 }
