+++ src/components/common/NavigableMonthHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as NavigableMonthHeader } from "./NavigavleMonthHeader.tsx";
